Patch to allow build under windows CI

+++ Inc/orbtraceIf.h
@@ -72,7 +72,7 @@ struct OrbtraceIfDevice
 
 struct __attribute__( ( packed ) ) dataBlock
 {
-            ssize_t fillLevel;                                   /* How full this block is */
+            long unsigned int fillLevel;                         /* How full this block is */
             uint8_t buffer[USB_TRANSFER_SIZE];                   /* Block buffer */
             struct libusb_transfer *usbtfr;                      /* USB Transfer handle */
 };
